+++ Source/Checker.cpp
@@ -383,14 +383,14 @@ bool face::Check(SGM::Result              &rResult,
 
     std::vector<SGM::Point3D> const &aPoints=GetPoints3D(rResult);
     std::vector<SGM::UnitVector3D> const &aNormals=GetNormals(rResult);
-    std::vector<size_t> const &aTriangles=GetTriangles(rResult);
+    std::vector<unsigned int> const &aTriangles=GetTriangles(rResult);
     size_t nTriangles=aTriangles.size();
     double dMaxAngle=0;
     for(Index1=0;Index1<nTriangles;Index1+=3)
         {
-        size_t a=aTriangles[Index1];
-        size_t b=aTriangles[Index1+1];
-        size_t c=aTriangles[Index1+2];
+        unsigned int a=aTriangles[Index1];
+        unsigned int b=aTriangles[Index1+1];
+        unsigned int c=aTriangles[Index1+2];
         SGM::Point3D const &A=aPoints[a];
         SGM::Point3D const &B=aPoints[b];
         SGM::Point3D const &C=aPoints[c];
